+++ .vercel/project.json
@@ -1,0 +1,1 @@
+{"neverMindDeployCard":true}
