+++ apps/docs/components/blocks.tsx
@@ -1,3 +1,3 @@
 'use client';
 
-export * from '@/example-library/blocks';
+export * from '@/rnr-ui/blocks';
