+++ apps/docs/content/docs/blocks/example-block/example.mdx
@@ -50,7 +50,7 @@ import { MultipleButtons } from '@docs/components/blocks';
 
         ```json doc-gen:file
         {
-          "file": "./node_modules/@rnr/example-library/reusables/components/ui/card.tsx",
+  "file": "./node_modules/@rnr/rnr-ui/reusables/components/ui/card.tsx",
           "codeblock": {
             "lang": "tsx",
             "meta": "title=\"@/components/ui/card.tsx\""
@@ -60,7 +60,7 @@ import { MultipleButtons } from '@docs/components/blocks';
 
         ```json doc-gen:file
         {
-          "file": "./node_modules/@rnr/example-library/src/blocks/multiple-buttons.tsx",
+  "file": "./node_modules/@rnr/rnr-ui/src/blocks/multiple-buttons.tsx",
           "codeblock": {
             "lang": "tsx",
             "meta": "title=\"@/components/blocks/multiple-buttons.tsx\""
