+++ apps/docs/content/docs/components/example-component.mdx
@@ -27,7 +27,7 @@ import { RedButtonPreview, RedButtonSecondaryPreview } from '@docs/components/ex
   <TabsContent value="code">
     ```json doc-gen:file
     {
-      "file": "./node_modules/@rnr/example-library/src/examples/red-button/red-button-primary.tsx",
+  "file": "./node_modules/@rnr/rnr-ui/src/examples/red-button/red-button-primary.tsx",
       "codeblock": {
         "lang": "tsx"
       }
@@ -61,7 +61,7 @@ import { RedButtonPreview, RedButtonSecondaryPreview } from '@docs/components/ex
 
         ```json doc-gen:file
         {
-          "file": "./node_modules/@rnr/example-library/reusables/components/ui/text.tsx",
+  "file": "./node_modules/@rnr/rnr-ui/reusables/components/ui/text.tsx",
           "codeblock": {
             "lang": "tsx",
             "meta": "title=\"@/components/ui/text.tsx\""
@@ -71,7 +71,7 @@ import { RedButtonPreview, RedButtonSecondaryPreview } from '@docs/components/ex
 
         ```json doc-gen:file
         {
-          "file": "./node_modules/@rnr/example-library/reusables/components/ui/button.tsx",
+  "file": "./node_modules/@rnr/rnr-ui/reusables/components/ui/button.tsx",
           "codeblock": {
             "lang": "tsx",
             "meta": "title=\"@/components/ui/button.tsx\""
@@ -81,7 +81,7 @@ import { RedButtonPreview, RedButtonSecondaryPreview } from '@docs/components/ex
 
         ```json doc-gen:file
         {
-          "file": "./node_modules/@rnr/example-library/src/components/ui/red-button.tsx",
+  "file": "./node_modules/@rnr/rnr-ui/src/components/ui/red-button.tsx",
           "codeblock": {
             "lang": "tsx",
             "meta": "title=\"@/components/ui/red-button.tsx\""
