+++ apps/docs/content/docs/components/form.mdx
@@ -27,7 +27,7 @@ import { FormBasicPreview, FormWithValidationPreview } from '@docs/components/ex
   <TabsContent value="code">
     ```json doc-gen:file
     {
-      "file": "./node_modules/@rnr/example-library/src/examples/form/form-basic.tsx",
+  "file": "./node_modules/@rnr/rnr-ui/src/examples/form/form-basic.tsx",
       "codeblock": {
         "lang": "tsx"
       }
@@ -61,7 +61,7 @@ import { FormBasicPreview, FormWithValidationPreview } from '@docs/components/ex
 
         ```json doc-gen:file
         {
-          "file": "./node_modules/@rnr/example-library/reusables/components/ui/text.tsx",
+  "file": "./node_modules/@rnr/rnr-ui/reusables/components/ui/text.tsx",
           "codeblock": {
             "lang": "tsx",
             "meta": "title=\"@/components/ui/text.tsx\""
@@ -71,7 +71,7 @@ import { FormBasicPreview, FormWithValidationPreview } from '@docs/components/ex
 
         ```json doc-gen:file
         {
-          "file": "./node_modules/@rnr/example-library/reusables/components/ui/label.tsx",
+  "file": "./node_modules/@rnr/rnr-ui/reusables/components/ui/label.tsx",
           "codeblock": {
             "lang": "tsx",
             "meta": "title=\"@/components/ui/label.tsx\""
@@ -81,7 +81,7 @@ import { FormBasicPreview, FormWithValidationPreview } from '@docs/components/ex
 
         ```json doc-gen:file
         {
-          "file": "./node_modules/@rnr/example-library/src/components/ui/forms.tsx",
+  "file": "./node_modules/@rnr/rnr-ui/src/components/ui/forms.tsx",
           "codeblock": {
             "lang": "tsx",
             "meta": "title=\"@/components/ui/forms.tsx\""
@@ -133,7 +133,7 @@ const [form] = Form.useForm();
   <TabsContent value="code">
     ```json doc-gen:file
     {
-      "file": "./node_modules/@rnr/example-library/src/examples/form/form-with-validation.tsx",
+  "file": "./node_modules/@rnr/rnr-ui/src/examples/form/form-with-validation.tsx",
       "codeblock": {
         "lang": "tsx"
       }
