+++ apps/docs/package.json
@@ -14,7 +14,7 @@
     "@radix-ui/react-select": "^2.2.5",
     "@radix-ui/react-slot": "^1.2.3",
     "@radix-ui/react-tabs": "^1.1.12",
-    "@rnr/example-library": "workspace:*",
+    "@rnr/rnr-ui": "workspace:*",
     "@rnr/utils": "workspace:*",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
