+++ apps/docs/public/r/form.json
@@ -11,7 +11,7 @@
   ],
   "files": [
     {
-      "path": "./node_modules/@rnr/example-library/src/components/ui/forms.tsx",
+  "path": "./node_modules/@rnr/rnr-ui/src/components/ui/forms.tsx",
       "content": "import FormRc, { Field } from 'rc-field-form';\nimport { FieldProps } from 'rc-field-form/es/Field';\nimport React from 'react';\nimport { View } from 'react-native';\nimport { Label } from '~/components/ui/label';\nimport { Text } from '~/components/ui/text';\n\nfunction FormItem<T>({\n  name,\n  label,\n  children,\n  ...restProps\n}: FieldProps<T> & {\n  label?: string;\n}) {\n  const hasRequired = restProps.rules?.some((rule: any) => rule.required);\n  return (\n    <Field name={name} {...restProps}>\n      {(control, meta, form) => {\n        const hasError = meta.errors && meta.errors.length > 0;\n        const errorMessage = meta.errors?.[0];\n\n        // Handle function children - pass control, meta, form exactly like rc-field-form\n        const childNode =\n          typeof children === 'function'\n            ? children(control, meta, form)\n            : React.isValidElement(children)\n              ? React.cloneElement(\n                  children as React.ReactElement<any>,\n                  {\n                    ...control,\n                    // Map React Native TextInput props\n                    value: control.value ?? '',\n                    onChangeText: (text: string) => control.onChange(text),\n                    onBlur: control.onBlur,\n                    // Add error styling if needed\n                    className: hasError\n                      ? `${(children as any).props?.className || ''} border-destructive`.trim()\n                      : (children as any).props?.className,\n                  } as any\n                )\n              : children;\n\n        // Only add UI wrapper if label or error exists\n        if (!label && !errorMessage) {\n          return childNode;\n        }\n\n        return (\n          <View className=\"mb-4\">\n            {label && (\n              <Label className=\"mb-2 text-base font-medium\">\n                {label}\n                {hasRequired && <Text className=\"text-destructive ml-1\">*</Text>}\n              </Label>\n            )}\n            {childNode}\n            {errorMessage && (\n              <Text variant=\"small\" className=\"text-destructive mt-1\">\n                {errorMessage}\n              </Text>\n            )}\n          </View>\n        );\n      }}\n    </Field>\n  );\n}\n\n// Create Form component with Item attached\nconst Form = FormRc as typeof FormRc & {\n  Item: typeof FormItem;\n  useForm: typeof FormRc.useForm;\n};\n\nForm.Item = FormItem;\nForm.useForm = FormRc.useForm;\n\nexport { Form };\n",
       "type": "registry:component"
     }
