+++ apps/showcase/app/components/example-block.tsx
@@ -1,6 +1,6 @@
 import { PreviewCarousel } from '@showcase/components/preview-carousel';
 import * as React from 'react';
-import { MultipleButtons } from '@/example-library/blocks/multiple-buttons';
+import { MultipleButtons } from '@/rnr-ui/blocks/multiple-buttons';
 
 const exampleBlockPreviews = [{ name: 'Multiple Buttons', component: MultipleButtons }];
