+++ apps/showcase/public/r/example-block.json
@@ -12,8 +12,8 @@
   ],
   "files": [
     {
-      "path": "./node_modules/@rnr/example-library/src/blocks/multiple-buttons.tsx",
-      "content": "import { View } from 'react-native';\nimport { RedButton } from '@/example-library/components/ui/red-button';\nimport { Card, CardContent, CardDescription, CardHeader, CardTitle } from '~/components/ui/card';\n\nexport function MultipleButtons() {\n  return (\n    <View className=\"gap-6\">\n      <Card className=\"border-border/0 sm:border-border shadow-none sm:shadow-sm sm:shadow-black/5\">\n        <CardHeader>\n          <CardTitle className=\"text-center text-xl sm:text-left\">Multiple Buttons</CardTitle>\n          <CardDescription className=\"text-center sm:text-left\">\n            Multiple buttons with different variants\n          </CardDescription>\n        </CardHeader>\n        <CardContent className=\"gap-6\">\n          <View className=\"gap-6\">\n            <RedButton onPress={() => {}} label=\"Primary Red Button\" />\n            <RedButton variant=\"secondary\" onPress={() => {}} label=\"Secondary Red Button\" />\n            <RedButton variant=\"outline\" onPress={() => {}} label=\"Outline Red Button\" />\n            <RedButton variant=\"ghost\" onPress={() => {}} label=\"Ghost Red Button\" />\n            <RedButton variant=\"link\" onPress={() => {}} label=\"Red Button Link\" />\n          </View>\n        </CardContent>\n      </Card>\n    </View>\n  );\n}\n",
+  "path": "./node_modules/@rnr/rnr-ui/src/blocks/multiple-buttons.tsx",
+  "content": "import { View } from 'react-native';\nimport { RedButton } from '@/rnr-ui/components/ui/red-button';\nimport { Card, CardContent, CardDescription, CardHeader, CardTitle } from '~/components/ui/card';\n\nexport function MultipleButtons() {\n  return (\n    <View className=\"gap-6\">\n      <Card className=\"border-border/0 sm:border-border shadow-none sm:shadow-sm sm:shadow-black/5\">\n        <CardHeader>\n          <CardTitle className=\"text-center text-xl sm:text-left\">Multiple Buttons</CardTitle>\n          <CardDescription className=\"text-center sm:text-left\">\n            Multiple buttons with different variants\n          </CardDescription>\n        </CardHeader>\n        <CardContent className=\"gap-6\">\n          <View className=\"gap-6\">\n            <RedButton onPress={() => {}} label=\"Primary Red Button\" />\n            <RedButton variant=\"secondary\" onPress={() => {}} label=\"Secondary Red Button\" />\n            <RedButton variant=\"outline\" onPress={() => {}} label=\"Outline Red Button\" />\n            <RedButton variant=\"ghost\" onPress={() => {}} label=\"Ghost Red Button\" />\n            <RedButton variant=\"link\" onPress={() => {}} label=\"Red Button Link\" />\n          </View>\n        </CardContent>\n      </Card>\n    </View>\n  );\n}\n",
       "type": "registry:block"
     }
   ]
