+++ packages/rnr-ui/src/examples/red-button/red-button-primary.tsx
@@ -1,4 +1,4 @@
-import { RedButton } from '@/example-library/components/ui/red-button';
+import { RedButton } from '@/rnr-ui/components/ui/red-button';
 
 export function RedButtonPreview() {
   return <RedButton label="Red Button" />;
