+++ packages/rnr-ui/src/examples/red-button/red-button-secondary.tsx
@@ -1,4 +1,4 @@
-import { RedButton } from '@/example-library/components/ui/red-button';
+import { RedButton } from '@/rnr-ui/components/ui/red-button';
 
 export function RedButtonSecondaryPreview() {
   return <RedButton label="Secondary Red Button" variant="secondary" />;
